File architecture refactoring.
